+++ example/c/include/ErrorEnum.h
