+++ example/c/include/ErrorStruct.h
