+++ example/c/include/MyStruct.h
