+++ example/c/include/Opaque.h
