+++ example/c/include/ResultOpaque.h
