+++ example/c/include/diplomat_runtime.h
