+++ example/c/include/result_ErrorEnum_box_ResultOpaque.h
