+++ example/c/include/result_box_ResultOpaque_ErrorEnum.h
