+++ example/c/include/result_box_ResultOpaque_void.h
