+++ example/c/include/result_void_box_ResultOpaque.h
