+++ example/c/include/result_void_void.h
